+++ apps/desktop/src/main/const/dir.ts
@@ -31,13 +31,6 @@ export const legacyLocalDbDir = path.join(appStorageDir, 'lobehub-local-db');
 export const FILE_STORAGE_DIR = 'file-storage';
 // Plugin installation directory
 export const INSTALL_PLUGINS_DIR = 'plugins';
-// Heterogeneous-agent (CC / Codex) working directory — holds the downloaded
-// file cache (`<HETERO_AGENT_DIR>/files`) and, in packaged builds, CLI trace
-// sessions (`<HETERO_AGENT_DIR>/tracing`). Single source of truth so the
-// controller and the Help-menu "open dir" entry never drift.
-export const HETERO_AGENT_DIR = 'heteroAgent';
-export const HETERO_AGENT_FILES_DIR = `${HETERO_AGENT_DIR}/files`;
-export const HETERO_AGENT_TRACING_DIR = `${HETERO_AGENT_DIR}/tracing`;
 
 // Desktop file service
 export const LOCAL_STORAGE_URL_PREFIX = '/lobe-desktop-file';
